Style Fix1

+++ modules/longest_common_subsequence/src/longest_common_subsequence_app.cpp
@@ -1,8 +1,9 @@
 // Copyright 2023 Kudryashov Nikita
 
-#include "include/longest_common_subsequence.h";
+#include "include/longest_common_subsequence.h"
 #include "include/longest_common_subsequence_app.h"
 #include<sstream>
+#include<cstring>
 
 LongestCommonSubsequenceApp::LongestCommonSubsequenceApp() : message_("") {}
 
@@ -80,7 +81,7 @@ std::string LongestCommonSubsequenceApp::operator()(int argc,
 
     std::ostringstream stream;
 
-    for (int i = 0; i < res.size(); i++) {
+    for (long unsigned int i = 0; i < res.size(); i++) {
         stream << res[i] << " ";
     }
